LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.8_Literals/7.8.4_String_Literals/S7.8.4_A7.1_T4-expected.txt

CONSOLE MESSAGE: line 77: SyntaxError: Unrecognized token '"\u'
S7.8.4_A7.1_T4

PASS Expected parsing failure

TEST COMPLETE
